Transform image to data uri if it isn't already
